requirements.txt
numpy
PyYAML
matplotlib
tqdm
pygame